List of fuctions

PGopt.particle_Gibbs — Method

particle_Gibbs(u_training, y_training, K, K_b, k_d, N, phi::Function, Lambda_Q, ell_Q, Q_init, V, A_init, x_init_dist, g, R; x_prim=nothing)

Run particle Gibbs sampler with ancestor sampling to obtain samples $\{A, Q, x_{T:-1}\}^{[1:K]}$ from the joint parameter and state posterior distribution $p(A, Q, x_{T:-1} \mid \mathbb{D}=\{u_{T:-1}, y_{T:-1}\})$.

Arguments

u_training: training input trajectory

y_training: training output trajectory

K: number of models/scenarios to be sampled

K_b: length of the burn in period

k_d: number of models/scenarios to be skipped to decrease correlation (thinning)

N: number of particles

phi: basis functions

Lambda_Q: scale matrix of IW prior on $Q$

ell_Q: degrees of freedom of IW prior on $Q$

Q_init: initial value of $Q$

V: left covariance matrix of MN prior on $A$

A_init: initial value of $A$

x_init_dist: distribution of the initial state

g: observation function

R: variance of zero-mean Gaussian measurement noise

x_prim: prespecified trajectory for the first iteration

This function is based on the papers

A. Svensson and T. B. Schön, “A flexible state–space model for learning nonlinear dynamical systems,” Automatica, vol. 80, pp. 189–199, 2017.

F. Lindsten, T. B. Schön, and M. Jordan, “Ancestor sampling for particle Gibbs,” Advances in Neural Information Processing Systems, vol. 25, 2012.

and the code provided in the supplementary material.

PGopt.solve_PG_OCP — Method

solve_PG_OCP(PG_samples::Vector{PG_sample}, phi::Function, R, H, u_min, u_max, y_min, y_max, R_cost_diag; x_vec_0=nothing, v_vec=nothing, e_vec=nothing, u_init=nothing, K_pre_solve=0, active_constraints=nothing, opts=nothing, print_progress=true)

Solve the optimal control problem of the following form:

$\min \sum_{t=0}^{H} \frac{1}{2} u_t \operatorname{diag}(R_{\mathrm{cost}}) u_t$

subject to:

\[\begin{aligned} \forall k, \forall t \\ x_{t+1}^{[k]} &= f_{\theta^{[k]}}(x_t^{[k]}, u_t) + v_t^{[k]} \\ x_{t, 1:n_y}^{[k]} &\geq y_{\mathrm{min},\ t} - e_t^{[k]} \\ x_{t, 1:n_y}^{[k]} &\leq y_{\mathrm{max},\ t} - e_t^{[k]} \\ u_t &\geq u_{\mathrm{min},\ t} \\ u_t &\leq u_{\mathrm{max},\ t}. \end{aligned}\]

Note that the output constraints imply the measurement function $y_t^{[k]} = x_{t, 1:n_y}^{[k]}$. Further note that the states of the individual models ($x^{[1:K]}$) are combined in the vector x_vec of dimension K * n_x.

Arguments

PG_samples: PG samples

phi: basis functions

R: variance of zero-mean Gaussian measurement noise - only used if e_vec is not passed

H: horizon of the OCP

u_min: array of dimension 1 x 1 or 1 x H containing the minimum control input for all timesteps

u_max: array of dimension 1 x 1 or 1 x H containing the maximum control input for all timesteps

y_min: array of dimension 1 x 1 or 1 x H containing the minimum system output for all timesteps

y_max: array of dimension 1 x 1 or 1 x H containing the maximum system output for all timesteps

R_cost_diag: parameter of the diagonal quadratic cost function

x_vec_0: vector with $K n_x$ elements containing the initial state of all models - if not provided, the initial states are sampled based on the PGS samples

v_vec: array of dimension n_x x H x K that contains the process noise for all models and all timesteps - if not provided, the noise is sampled based on the PGS samples

e_vec: array of dimension n_y x H x K that contains the measurement noise for all models and all timesteps - if not provided, the noise is sampled based on the provided R

u_init: initial guess for the optimal trajectory

K_pre_solve: if K_pre_solve > 0, an initial guess for the optimal trajectory is obtained by solving the OCP with only K_pre_solve < K models

active_constraints: vector containing the indices of the models, for which the output constraints are active - if not provided, the output constraints are considered for all models

opts: SolverOptions struct containing options of the solver

print_progress: if set to true, the progress is printed

PGopt.solve_PG_OCP_greedy_guarantees — Method

solve_PG_OCP_greedy_guarantees(PG_samples::Vector{PG_sample}, phi::Function, R, H, u_min, u_max, y_min, y_max, R_cost_diag, β; x_vec_0=nothing, v_vec=nothing, e_vec=nothing, u_init=nothing, K_pre_solve=0, opts=nothing, print_progress=true)

Solve the following optimal control problem and determine a support sub-sample with cardinality $s$ via a greedy constraint removal. Based on the cardinality $s$, a bound on the probability that the incurred cost exceeds the worst-case cost or that the constraints are violated when the input trajectory $u_{0:H}$ is applied to the unknown system is calculated (i.e., $1-\epsilon$ is determined).

$\min \sum_{t=0}^{H} \frac{1}{2} u_t \operatorname{diag}(R_{\mathrm{cost}}) u_t$

subject to:

\[\begin{aligned} \forall k, \forall t \\ x_{t+1}^{[k]} &= f_{\theta^{[k]}}(x_t^{[k]}, u_t) + v_t^{[k]} \\ x_{t, 1:n_y}^{[k]} &\geq y_{\mathrm{min},\ t} - e_t^{[k]} \\ x_{t, 1:n_y}^{[k]} &\leq y_{\mathrm{max},\ t} - e_t^{[k]} \\ u_t &\geq u_{\mathrm{min},\ t} \\ u_t &\leq u_{\mathrm{max},\ t}. \end{aligned}\]

Note that the output constraints imply the measurement function $y_t^{[k]} = x_{t, 1:n_y}^{[k]}$. Further note that the states of the individual models ($x^{[1:K]}$) are combined in the vector x_vec of dimension K * n_x.

Arguments

PG_samples: PG samples

phi: basis functions

R: variance of zero-mean Gaussian measurement noise - only used if e_vec is not passed

H: horizon of the OCP

u_min: array of dimension 1 x 1 or 1 x H containing the minimum control input for all timesteps

u_max: array of dimension 1 x 1 or 1 x H containing the maximum control input for all timesteps

y_min: array of dimension 1 x 1 or 1 x H containing the minimum system output for all timesteps

y_max: array of dimension 1 x 1 or 1 x H containing the maximum system output for all timesteps

R_cost_diag: parameter of the diagonal quadratic cost function

β: confidence parameter

x_vec_0: vector with K*n_x elements containing the initial state of all models - if not provided, the initial states are sampled based on the PGS samples

v_vec: array of dimension n_x x H x K that contains the process noise for all models and all timesteps - if not provided, the noise is sampled based on the PGS samples

e_vec: array of dimension n_y x H x K that contains the measurement noise for all models and all timesteps - if not provided, the noise is sampled based on the provided R

u_init: initial guess for the optimal trajectory

K_pre_solve: if K_pre_solve > 0, an initial guess for the optimal trajectory is obtained by solving the OCP with only K_pre_solve < K models

opts: SolverOptions struct containing options of the solver

print_progress: if set to true, the progress is printed

PGopt.test_prediction — Method

test_prediction(PG_samples::Vector{PG_sample}, phi::Function, g, R, k_n, u_test, y_test)

Simulate the PGS samples forward in time and compare the predictions to the test data.

Arguments

PG_samples: PG samples

phi: basis functions

g: observation function

R: variance of zero-mean Gaussian measurement noise

k_n: each model is simulated $k_n$ times

u_test: test input

y_test: test output

PGopt.plot_predictions — Method

plot_predictions(y_pred, y_test; plot_percentiles=false, y_min=nothing, y_max=nothing)

Plot the predictions and the test data.

Arguments

y_pred: matrix containing the output predictions

y_test: test output trajectory

plot_percentiles: if set to true, percentiles are plotted

y_min: min output to be plotted as constraint

y_max: max output to be plotted as constraint

PGopt.plot_autocorrelation — Method

plot_autocorrelation(PG_samples::Vector{PG_sample}; max_lag=0)

Plot the autocorrelation function (ACF) of the PG samples. This might be helpful when adjusting the thinning parameter $k_d$.

Arguments

PG_samples: PG samples

max_lag: maximum lag at which to calculate the ACF

PGopt.epsilon — Method

epsilon(s::Int64, K::Int64, β::Float64)

Determine the parameter $\epsilon$. $1-\epsilon$ corresponds to a bound on the probability that the incurred cost exceeds the worst-case cost or that the constraints are violated when the input trajectory $u_{0:H}$ is applied to the unknown system. $\epsilon$ is the unique solution over the interval $(0,1)$ of the polynomial equation in the $v$ variable:

$\binom{K}{s}(1-v)^{K-s}-\frac{\beta}{K}\sum_{m=s}^{K-1}\binom{m}{s}(1-v)^{m-s}=0$.

Arguments

s: cardinality of the support sub-sample

K: number of scenarios

β: confidence parameter

This function is based on the paper

S. Garatti and M. C. Campi, “Risk and complexity in scenario optimization,” Mathematical Programming, vol. 191, no. 1, pp. 243–279, 2022.

and the code provided in the appendix.

PGopt.MNIW_sample — Method

MNIW_sample(Phi, Psi, Sigma, V, Lambda_Q, ell_Q, T)

Sample new model parameters $\{A, Q\}$ from the conditional distribution $p(A, Q | x_{T:-1})$, which is a matrix normal inverse Wishart (MNIW) distribution.

Arguments

Phi: statistic; see paper below for definition

Psi: statistic; see paper below for definition

Sigma: statistic; see paper below for definition

V: left covariance matrix of MN prior on $A$

Lambda_Q: scale matrix of IW prior on $Q$

ell_Q: degrees of freedom of IW prior on $Q$

T: length of the training trajectory

This function is based on the paper

A. Svensson and T. B. Schön, “A flexible state–space model for learning nonlinear dynamical systems,” Automatica, vol. 80, pp. 189–199, 2017.

and the code provided in the supplementary material.

PGopt.systematic_resampling — Method

systematic_resampling(W, N)

Sample N indices according to the weights W.

Arguments

W: vector containing the weights of the particles

N: number of indices to be sampled

This function is based on the paper

A. Svensson and T. B. Schön, “A flexible state–space model for learning nonlinear dynamical systems,” Automatica, vol. 80, pp. 189–199, 2017.

and the code provided in the supplementary material.